third_party/WebKit/LayoutTests/bluetooth/descriptor/writeValue/write-updates-value.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('DisconnectingHealthThermometerAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('health_thermometer'))
    .then(service => service.getCharacteristic('measurement_interval'))
    .then(characteristic => characteristic.getDescriptor(user_description.name))
    .then(descriptor => {
      assert_equals(descriptor.value, null);
      let textEncoder = new TextEncoder();
      let newValue = textEncoder.encode('foo');
      assert_true(newValue instanceof Uint8Array, 'newValue is Uint8Array');
      return descriptor.writeValue(newValue).then(() => {
        assert_true(descriptor.value instanceof DataView, 'descriptor.value is DataView');
        assert_array_equals(Array.from(descriptor.value.getUint8()), Array.from(newValue.buffer));
      })
    })
}, 'Succesful read should update descriptor\'s value.');
</script>
